// src/api/generated/model/imageGroupRead.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * DentC Backend
 * Dental PMS REST API. Column-tenant-scoped; snake_case; Orval-ready.
 * OpenAPI spec version: 1.0.0
 */

export interface ImageGroupRead {
  tenant_id: number;
  office_id?: number | null;
  patient_id?: number | null;
  template_id?: number | null;
  legacy_id?: string | null;
  name?: string | null;
  group_type?: string | null;
  ext_id?: string | null;
  source?: string | null;
  source2?: string | null;
  is_deleted: boolean;
  created_by?: string | null;
  id: number;
  created_at: string;
  updated_at?: string | null;
}